refactor: adjust screens logic
