+++ products.test.json
@@ -1,0 +1,1 @@
+[{"description":"BACON LOMBO KASSEL kg","code":"29617","quantity":0,"unity_measure":" KG","unitary_value":45.9},{"description":"PIMENTAO VERMELHO kg","code":"885","quantity":0,"unity_measure":" KG","unitary_value":12.99},{"description":"PIMENTAO VERDE kg","code":"882","quantity":0,"unity_measure":" KG","unitary_value":6.99},{"description":"AMENDOIM TORRADO S/PELE S/SAL kg","code":"20963","quantity":0,"unity_measure":" KG","unitary_value":29.9},{"description":"CHAMPIGNONS FAT GOMES DA COSTA 170G","code":"20375","quantity":1,"unity_measure":" UN","unitary_value":9.9},{"description":"BATATA LAVADA kg","code":"827","quantity":1,"unity_measure":" KG","unitary_value":5.99},{"description":"GENGIBRE kg","code":"759","quantity":0,"unity_measure":" KG","unitary_value":13.99},{"description":"CEBOLINHA ORG SOLO VIVO 50G","code":"55547","quantity":1,"unity_measure":" UN","unitary_value":7.99},{"description":"ACUCAR UNIAO 1kg","code":"3409","quantity":1,"unity_measure":" UN","unitary_value":2.99},{"description":"AZEITE ESPORAO HERDADE 500ML","code":"9202","quantity":1,"unity_measure":" UN","unitary_value":42.9},{"description":"MOLHO SHOYU SAKURA TRAD 150ML","code":"6832","quantity":1,"unity_measure":" UN","unitary_value":3.29},{"description":"CALDO KNORR 57G GALINHA","code":"4716","quantity":1,"unity_measure":" UN","unitary_value":1.89}]
